fit for xuperunion v3.1
