+++ conf/gen.json
@@ -1,0 +1,13 @@
+{
+    "type": "xchain",
+    "workNum": 2,
+    "mode": "local",
+	"host": "127.0.0.1:37101",
+	"chain": "xuper",
+    "rounds": [
+        {
+            "label": "generate",
+            "number": [ 10 ]
+        }
+    ]
+}
