+++ conf/plugins.conf
@@ -1,0 +1,8 @@
+{
+	"crypto":[{
+		"subtype": "default",
+			"path": "plugins/crypto/crypto-default.so.1.0.0",
+			"version": "1.0.0",
+			"ondemand": false
+	}]
+}
