+++ conf/remote.json
@@ -2,8 +2,8 @@
     "type": "demo",
     "workNum": 5,
     "mode": "remote",
-	"host": "192.168.1.1:37101",
-    "broker":"192.168.1.1:6379",
+	"host": "127.0.0.1:37101",
+    "broker":"127.0.0.1:6379",
     "resultBackend":"BackendChan",
     "pubSubChan":"PubSubChan",
     "rounds": [
