+++ data/keys/public.key
@@ -1,0 +1,1 @@
+{"Curvname":"P-256","X":74695617477160058757747208220371236837474210247114418775262229497812962582435,"Y":51348715319124770392993866417088542497927816017012182211244120852620959209571}
