+++ monitor/facility.go
@@ -2,6 +2,8 @@ package monitor
 
 import (
 	"encoding/json"
+	"encoding/gob"
+	"bytes"
 	"fmt"
 	"time"
 	"github.com/xuperchain/xuperbench/log"
@@ -23,14 +25,15 @@ type TpsInfo struct {
 func TpsSummary(testStats []*LatencyStats) []TpsInfo {
 	ret := make([]TpsInfo, 0, len(testStats))
 	for i, v := range testStats {
+		fmt.Println(v)
 		tps := TpsInfo{
 			Round:    i + 1,
-			Name:     v.ext,
-			Succ:     v.succ,
-			Fail:     v.fail,
+			Name:     v.Ext,
+			Succ:     v.Succ,
+			Fail:     v.Fail,
 			SendRate: fmt.Sprintf("%d", v.RatePerSec()),
-			MaxLat:   v.max,
-			MinLat:   v.min,
+			MaxLat:   v.Max,
+			MinLat:   v.Min,
 			AvgLat:   v.Avg(),
 			Tps:      fmt.Sprintf("%v tps", v.TPS()),
 		}
@@ -46,8 +49,8 @@ func PrintFormat(testStats []*LatencyStats) {
 	table.AddHeaders("Round", "Name", "Succ", "Fail", "Send Rate",
 		"Max Latency", "Min Latency", "Avg Latency", "Throughput")
 	for i, v := range testStats {
-		table.AddRow(i+1, v.ext, v.succ, v.fail, fmt.Sprintf("%d", v.RatePerSec()),
-			v.max, v.min, v.Avg(), fmt.Sprintf("%v tps", v.TPS()))
+		table.AddRow(i+1, v.Ext, v.Succ, v.Fail, fmt.Sprintf("%d", v.RatePerSec()),
+			v.Max, v.Min, v.Avg(), fmt.Sprintf("%v tps", v.TPS()))
 	}
 
 	log.INFO.Println("benchmark summary...\n" + table.Render())
@@ -67,15 +70,15 @@ func Marshal(gStat []*LatencyStats) string {
 	for i, v := range gStat {
 		m := make(map[string]interface{})
 		m["RoundId"] = i + 1
-		m["Name"] = v.ext
-		m["Succ"] = v.succ
-		m["Fail"] = v.fail
-		m["MaxLat"] = v.max
-		m["MinLat"] = v.min
+		m["Name"] = v.Ext
+		m["Succ"] = v.Succ
+		m["Fail"] = v.Fail
+		m["MaxLat"] = v.Max
+		m["MinLat"] = v.Min
 		m["AvgLat"] = v.Avg()
-		m["TotLat"] = v.dur
-		m["Begin"] = v.begin.UnixNano()
-		m["End"] = v.end.UnixNano()
+		m["TotLat"] = v.Dur
+		m["Begin"] = v.Begin.UnixNano()
+		m["End"] = v.End.UnixNano()
 		m["Sendrate"] = v.RatePerSec()
 
 		ret = append(ret, m)
@@ -88,4 +91,32 @@ func Marshal(gStat []*LatencyStats) string {
 	}
 
 	return string(data)
+}
+
+func Serialize(gStat []*LatencyStats) string {
+	data := []LatencyStats{}
+	for _, v := range gStat {
+		data = append(data, *v)
+	}
+	fmt.Println(data)
+	var buf bytes.Buffer
+	enc := gob.NewEncoder(&buf)
+	err := enc.Encode(data)
+	if err != nil {
+		log.ERROR.Printf("encount error <%s>", err)
+		return ""
+	}
+	return buf.String()
+}
+
+func Deserialize(str string) []LatencyStats {
+	var stat []LatencyStats
+	buf := bytes.NewBufferString(str)
+	dec := gob.NewDecoder(buf)
+	err := dec.Decode(&stat)
+	if err != nil {
+		log.ERROR.Printf("encount error <%s>", err)
+		return nil
+	}
+	return stat
 }
