+++ monitor/stat.go
@@ -10,126 +10,126 @@ import (
 const NORMAL_REQ = 500
 
 type LatencyStats struct {
-	// 增加begin，end信息，便于计算压测速率信息
-	// begin： 压测开始时间
-	// end：压测结束时间
-	begin, end time.Time
+	// 增加Begin，End信息，便于计算压测速率信息
+	// Begin： 压测开始时间
+	// End：压测结束时间
+	Begin, End time.Time
 
 	// 统计成功次数，失败次数
-	succ, fail int
+	Succ, Fail int
 
 	// 总压测次数
-	num int
+	Num int
 
-	mark time.Time
+	Mark time.Time
 
 	// 总时延，最小时延，最大时延
-	dur, min, max time.Duration
+	Dur, Min, Max time.Duration
 
 	// 互斥锁，保证互操作时 goroutine 并发安全
 	mu *sync.Mutex
 
 	// 附加信息，便于最后统计汇总
-	ext string
+	Ext string
 
-	percentile [NORMAL_REQ]int
+	Percentile [NORMAL_REQ]int
 
-	longreq []int
+	Longreq []int
 }
 
-func New(ext string) *LatencyStats {
+func New(Ext string) *LatencyStats {
 	s := &LatencyStats{
-		ext: ext,
+		Ext: Ext,
 		mu:  &sync.Mutex{},
-		percentile: [NORMAL_REQ]int{},
-		longreq: []int{},
+		Percentile: [NORMAL_REQ]int{},
+		Longreq: []int{},
 	}
 	return s
 }
 
 func (s *LatencyStats) Start() {
-	s.mark = time.Now()
-	if s.begin.IsZero() {
-		s.begin = s.mark
+	s.Mark = time.Now()
+	if s.Begin.IsZero() {
+		s.Begin = s.Mark
 	}
 }
 
 func (s LatencyStats) String() string {
-	return fmt.Sprintf("begin: %v, end: %v, succ: %v, fail :%v, max-latency: %v, min-latency: %v",
-		s.begin.UnixNano(), s.end.UnixNano(), s.succ, s.fail, s.max, s.fail)
+	return fmt.Sprintf("Begin: %v, End: %v, Succ: %v, Fail :%v, max-latency: %v, min-latency: %v",
+		s.Begin.UnixNano(), s.End.UnixNano(), s.Succ, s.Fail, s.Max, s.Fail)
 }
 
 func (s *LatencyStats) record(n int) {
 
-	if s.mark.IsZero() {
+	if s.Mark.IsZero() {
 		panic("not started")
 	}
 
 	now := time.Now()
-	s.end = now
-	dur := now.Sub(s.mark)
+	s.End = now
+	dur := now.Sub(s.Mark)
 	dur1 := dur / time.Duration(n)
-	if dur1 < s.min || s.min == 0 {
-		s.min = dur1
+	if dur1 < s.Min || s.Min == 0 {
+		s.Min = dur1
 	}
-	if dur1 > s.max {
-		s.max = dur1
+	if dur1 > s.Max {
+		s.Max = dur1
 	}
-	s.dur += dur
-	s.num += n
+	s.Dur += dur
+	s.Num += n
 	ms := int(dur1 / time.Millisecond)
 	if ms < NORMAL_REQ {
-		s.percentile[ms]++
+		s.Percentile[ms]++
 	} else {
-		s.longreq = append(s.longreq, ms)
+		s.Longreq = append(s.Longreq, ms)
 	}
-	s.mark = time.Time{}
+	s.Mark = time.Time{}
 }
 
 func (s *LatencyStats) RecordSucc(n int) {
 	s.record(n)
-	s.succ += n
+	s.Succ += n
 }
 
 func (s *LatencyStats) RecordFail(n int) {
 	s.record(n)
-	s.fail += n
+	s.Fail += n
 }
 
 func (s *LatencyStats) RatePerSec() int {
-	durSec := s.dur / time.Second
+	durSec := s.Dur / time.Second
 	if durSec > 0 {
-		return s.num / int(durSec)
+		return s.Num / int(durSec)
 	}
-	return s.num
+	return s.Num
 }
 
 func (s *LatencyStats) Avg() time.Duration {
-	if s.num > 0 {
-		return s.dur / time.Duration(s.num)
+	if s.Num > 0 {
+		return s.Dur / time.Duration(s.Num)
 	}
 	return 0
 }
 
 func (s *LatencyStats) TPS() int {
-	lat := s.end.Sub(s.begin)
-	return int(float64(s.succ) / float64(lat) * float64(time.Second))
+	lat := s.End.Sub(s.Begin)
+	return int(float64(s.Succ) / float64(lat) * float64(time.Second))
 }
 
 func (s *LatencyStats) Pct(p int) string {
-	n := s.num * p / 100
+	n := s.Num * p / 100
 	i := 0
 	t := 0
 	for i < NORMAL_REQ && t < n {
-		t += s.percentile[i]
+		t += s.Percentile[i]
 		i += 1
 	}
 	if t >= n {
 		return fmt.Sprintf("< %dms", i + 1)
 	} else {
-		sort.Ints(s.longreq)
-		if len(s.longreq) > n - t {
-			return fmt.Sprintf("< %dms", s.longreq[n - t] + 1)
+		sort.Ints(s.Longreq)
+		if len(s.Longreq) > n - t {
+			return fmt.Sprintf("< %dms", s.Longreq[n - t] + 1)
 		} else {
 			return "NaN"
 		}
@@ -142,32 +142,32 @@ func (s *LatencyStats) Add(x *LatencyStats) {
 		s.mu.Unlock()
 	}()
 
-	if s.ext == "" {
-		s.ext = x.ext
+	if s.Ext == "" {
+		s.Ext = x.Ext
 	}
-	if s.begin.UnixNano() > x.begin.UnixNano() {
-		s.begin = x.begin
+	if s.Begin.IsZero()|| s.Begin.UnixNano() > x.Begin.UnixNano() {
+		s.Begin = x.Begin
 	}
-	if s.end.IsZero() || s.end.UnixNano() < x.end.UnixNano() {
-		s.end = x.end
+	if s.End.IsZero() || s.End.UnixNano() < x.End.UnixNano() {
+		s.End = x.End
 	}
 
-	if x.min < s.min || s.min == 0 {
-		s.min = x.min
+	if x.Min < s.Min || s.Min == 0 {
+		s.Min = x.Min
 	}
-	if x.max > s.max {
-		s.max = x.max
+	if x.Max > s.Max {
+		s.Max = x.Max
 	}
-	s.dur += x.dur
-	s.num += x.num
+	s.Dur += x.Dur
+	s.Num += x.Num
 
-	s.succ += x.succ
-	s.fail += x.fail
+	s.Succ += x.Succ
+	s.Fail += x.Fail
 
 	for i:=0; i<NORMAL_REQ; i++ {
-		s.percentile[i] += x.percentile[i]
+		s.Percentile[i] += x.Percentile[i]
 	}
-	s.longreq = append(s.longreq, x.longreq...)
+	s.Longreq = append(s.Longreq, x.Longreq...)
 }
 
 // Merge merge multiple LatencyStats(chan *LatencyStats) generated by goroutines and return one LatencyStats
